requirements.txt
cryptography==41.0.7
pycryptodome==3.19.0
ecdsa==0.18.0
numpy==1.24.3
py_ecc==6.0.0